refactor : 로또 당첨 여부를 판단하는 로직 변경
LottoRank Enum 안에 보너스 넘버 매칭 필요  유무를 판단하는 matchBonusNumber 필드 추가
