+++ src/test/java/lotto/util/RandomNumberGeneratorTest.java
@@ -1,6 +1,7 @@
-package lotto.domain.lottos.generator;
+package lotto.util;
 
 import lotto.domain.lottos.LottoNumber;
+import lotto.util.RandomNumberGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
